requirements.txt
groq
python-dotenv
gitpython
openai
